fix(meilisearch): move file and 0.3.3 compatible
