+++ heliotrope/server.py
@@ -28,7 +28,7 @@ from sentry_sdk.integrations.sanic import SanicIntegration
 
 from heliotrope import __version__
 from heliotrope.config import HeliotropeConfig
-from heliotrope.database.meilisearch.client import MeiliSearch
+from heliotrope.database.meilisearch import MeiliSearch
 from heliotrope.database.orm import ORM
 from heliotrope.interpreter import CommonJS
 from heliotrope.request.base import BaseRequest
